Move Imgui logic from ModuleUI to ModuleEditor
