+++ Engine/Module/ModuleInput.cpp
@@ -149,7 +149,7 @@ update_status ModuleInput::PreUpdate()
 
 		case SDL_DROPFILE:
 			char *dropped_filedir = event.drop.file;
-			App->ui->editor_ui->file_explorer_ui.CopyFileToSelectedFolder(dropped_filedir);
+			App->editor->editor_ui->file_explorer_ui.CopyFileToSelectedFolder(dropped_filedir);
 			SDL_free(dropped_filedir);
 			
 			break;
