+++ Engine/Module/ModuleRender.cpp
@@ -145,7 +145,7 @@ bool ModuleRender::CleanUp()
 void ModuleRender::Render() const
 {
 	BROFILER_CATEGORY("Global Render",Profiler::Color::Aqua);
-	App->ui->Render();
+	App->editor->Render();
 	BROFILER_CATEGORY("Swap Window (VSYNC)", Profiler::Color::Aquamarine);
 	SDL_GL_SwapWindow(App->window->window);
 	App->time->EndFrame();
